de-typscript deploys radiator and de-isolate it
- convert to ES6
- remove JSPM
- use webpack to build it
- use main package.json
