--- static/src/deploys-radiator/.gitignore
@@ -1,3 +1,0 @@
-/node_modules/
-/jspm_packages/
-/typings/
