--- static/src/deploys-radiator/config.js
@@ -1,115 +1,0 @@
-System.config({
-  baseURL: "/assets/admin/deploys-radiator",
-  defaultJSExtensions: true,
-  transpiler: "babel",
-  babelOptions: {
-    "optional": [
-      "runtime",
-      "optimisation.modules.system"
-    ]
-  },
-  typescriptOptions: {
-    "typeCheck": "strict",
-    "tsconfig": true
-  },
-  paths: {
-    "github:*": "jspm_packages/github/*",
-    "npm:*": "jspm_packages/npm/*"
-  },
-
-  packages: {
-    "app": {
-      "defaultExtension": "ts",
-      "meta": {
-        "*.ts": {
-          "loader": "ts"
-        }
-      }
-    }
-  },
-
-  map: {
-    "babel": "npm:babel-core@5.8.34",
-    "babel-runtime": "npm:babel-runtime@5.8.34",
-    "core-js": "npm:core-js@1.2.6",
-    "hyperscript-helpers": "npm:hyperscript-helpers@2.0.2",
-    "immutable": "npm:immutable@3.7.6",
-    "monapt": "npm:monapt@0.5.0",
-    "ts": "github:frankwallis/plugin-typescript@4.0.2",
-    "virtual-dom": "npm:virtual-dom@2.1.1",
-    "github:frankwallis/plugin-typescript@4.0.2": {
-      "typescript": "npm:typescript@1.8.9"
-    },
-    "github:jspm/nodelibs-assert@0.1.0": {
-      "assert": "npm:assert@1.3.0"
-    },
-    "github:jspm/nodelibs-os@0.1.0": {
-      "os-browserify": "npm:os-browserify@0.1.2"
-    },
-    "github:jspm/nodelibs-path@0.1.0": {
-      "path-browserify": "npm:path-browserify@0.0.0"
-    },
-    "github:jspm/nodelibs-process@0.1.2": {
-      "process": "npm:process@0.11.2"
-    },
-    "github:jspm/nodelibs-util@0.1.0": {
-      "util": "npm:util@0.10.3"
-    },
-    "npm:assert@1.3.0": {
-      "util": "npm:util@0.10.3"
-    },
-    "npm:babel-runtime@5.8.34": {
-      "process": "github:jspm/nodelibs-process@0.1.2"
-    },
-    "npm:core-js@1.2.6": {
-      "fs": "github:jspm/nodelibs-fs@0.1.2",
-      "path": "github:jspm/nodelibs-path@0.1.0",
-      "process": "github:jspm/nodelibs-process@0.1.2",
-      "systemjs-json": "github:systemjs/plugin-json@0.1.0"
-    },
-    "npm:error@4.4.0": {
-      "assert": "github:jspm/nodelibs-assert@0.1.0",
-      "camelize": "npm:camelize@1.0.0",
-      "string-template": "npm:string-template@0.2.1",
-      "xtend": "npm:xtend@4.0.1"
-    },
-    "npm:ev-store@7.0.0": {
-      "individual": "npm:individual@3.0.0"
-    },
-    "npm:global@4.3.0": {
-      "process": "npm:process@0.5.2"
-    },
-    "npm:inherits@2.0.1": {
-      "util": "github:jspm/nodelibs-util@0.1.0"
-    },
-    "npm:next-tick@0.2.2": {
-      "process": "github:jspm/nodelibs-process@0.1.2"
-    },
-    "npm:os-browserify@0.1.2": {
-      "os": "github:jspm/nodelibs-os@0.1.0"
-    },
-    "npm:path-browserify@0.0.0": {
-      "process": "github:jspm/nodelibs-process@0.1.2"
-    },
-    "npm:process@0.11.2": {
-      "assert": "github:jspm/nodelibs-assert@0.1.0"
-    },
-    "npm:typescript@1.8.9": {
-      "os": "github:jspm/nodelibs-os@0.1.0"
-    },
-    "npm:util@0.10.3": {
-      "inherits": "npm:inherits@2.0.1",
-      "process": "github:jspm/nodelibs-process@0.1.2"
-    },
-    "npm:virtual-dom@2.1.1": {
-      "browser-split": "npm:browser-split@0.0.1",
-      "error": "npm:error@4.4.0",
-      "ev-store": "npm:ev-store@7.0.0",
-      "global": "npm:global@4.3.0",
-      "is-object": "npm:is-object@1.0.1",
-      "next-tick": "npm:next-tick@0.2.2",
-      "x-is-array": "npm:x-is-array@0.1.0",
-      "x-is-string": "npm:x-is-string@0.1.0"
-    }
-  }
-});
