--- static/src/deploys-radiator/manual-typings/custom-window.d.ts
@@ -1,3 +1,0 @@
-interface Window {
-    gitHubAccessToken: string;
-}
