--- static/src/deploys-radiator/manual-typings/vdom-virtualize.d.ts
@@ -1,7 +1,0 @@
-/// <reference path="../typings/tsd.d.ts" />
-
-declare module 'vdom-virtualize' {
-    function main(node: Node): VirtualDOM.VNode;
-
-    export default main;
-}
