--- static/src/deploys-radiator/tsconfig.json
@@ -1,6 +1,0 @@
-{
-    "compilerOptions": {
-        "noImplicitAny": true,
-        "module": "system"
-    }
-}
