--- static/src/deploys-radiator/tsd.json
@@ -1,15 +1,0 @@
-{
-  "version": "v4",
-  "repo": "borisyankov/DefinitelyTyped",
-  "ref": "master",
-  "path": "typings",
-  "bundle": "typings/tsd.d.ts",
-  "installed": {
-    "virtual-dom/virtual-dom.d.ts": {
-      "commit": "4c41545030eba44b27ce7209112dbf862f6b1d78"
-    },
-    "whatwg-fetch/whatwg-fetch.d.ts": {
-      "commit": "ce8ae773688d2244ba11fdf3378fcd68f773e1c4"
-    }
-  }
-}
