+++ static/src/deploys-radiator/webpack.config.js
@@ -1,0 +1,12 @@
+module.exports = {
+    entry: `${__dirname}/app/main.js`,
+    output: {
+        path: 'static/target/deploys-radiator',
+        filename: 'main.js'
+    },
+    module: {
+        loaders: [
+            { test: /\.js$/, exclude: /node_modules/, loader: "babel-loader" }
+        ]
+    }
+};
